+++ tests/Unit/RoutesTest.php
@@ -147,7 +147,7 @@ class RoutesTest extends LaravelTableTestCase
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
         foreach ($users as $user) {
             $this->assertStringContainsString('edit-' . $user->id, $html);
-            $this->assertStringContainsString('action="http://localhost/users/edit?id=' . $user->id . '"', $html);
+            $this->assertStringContainsString('action="http://localhost/users/edit?' . $user->id . '"', $html);
         }
     }
 
@@ -161,7 +161,7 @@ class RoutesTest extends LaravelTableTestCase
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
         foreach ($users as $user) {
             $this->assertStringNotContainsString('<form class="edit-' . $user->id, $html);
-            $this->assertStringNotContainsString('action="http://localhost/users/edit?id=' . $user->id . '"', $html);
+            $this->assertStringNotContainsString('action="http://localhost/users/edit?' . $user->id . '"', $html);
         }
     }
 
@@ -178,7 +178,7 @@ class RoutesTest extends LaravelTableTestCase
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
         foreach ($users as $user) {
             $this->assertStringContainsString('destroy-' . $user->id, $html);
-            $this->assertStringContainsString('action="http://localhost/users/destroy?id=' . $user->id . '"', $html);
+            $this->assertStringContainsString('action="http://localhost/users/destroy?' . $user->id . '"', $html);
         }
     }
 
@@ -192,7 +192,7 @@ class RoutesTest extends LaravelTableTestCase
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
         foreach ($users as $user) {
             $this->assertStringNotContainsString('<form class="destroy-' . $user->id, $html);
-            $this->assertStringNotContainsString('action="http://localhost/users/destroy?id=' . $user->id . '"', $html);
+            $this->assertStringNotContainsString('action="http://localhost/users/destroy?' . $user->id . '"', $html);
         }
     }
 
@@ -209,7 +209,7 @@ class RoutesTest extends LaravelTableTestCase
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
         foreach ($users as $user) {
             $this->assertStringContainsString('show-' . $user->id, $html);
-            $this->assertStringContainsString('action="http://localhost/users/show?id=' . $user->id . '"', $html);
+            $this->assertStringContainsString('action="http://localhost/users/show?' . $user->id . '"', $html);
         }
     }
 
@@ -223,7 +223,7 @@ class RoutesTest extends LaravelTableTestCase
         $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
         foreach ($users as $user) {
             $this->assertStringNotContainsString('<form class="show-' . $user->id, $html);
-            $this->assertStringNotContainsString('action="http://localhost/users/show?id=' . $user->id . '"', $html);
+            $this->assertStringNotContainsString('action="http://localhost/users/show?' . $user->id . '"', $html);
         }
     }
 }