+++ tests/Unit/RowDisableTest.php
@@ -50,13 +50,13 @@ class RowDisableTest extends LaravelTableTestCase
             if ($user->id === 1 || $user->id === 2) {
                 $this->assertStringNotContainsString('edit-' . $user->id, $html);
                 $this->assertStringNotContainsString(
-                    'action="http://localhost/users/edit?id=' . $user->id . '"',
+                    'action="http://localhost/users/edit?' . $user->id . '"',
                     $html
                 );
             } else {
                 $this->assertStringContainsString('edit-' . $user->id, $html);
                 $this->assertStringContainsString(
-                    'action="http://localhost/users/edit?id=' . $user->id . '"',
+                    'action="http://localhost/users/edit?' . $user->id . '"',
                     $html
                 );
             }
@@ -93,12 +93,12 @@ class RowDisableTest extends LaravelTableTestCase
             if ($user->id === 1 || $user->id === 2) {
                 $this->assertStringNotContainsString('edit-' . $user->id, $html);
                 $this->assertStringNotContainsString(
-                    'action="http://localhost/users/edit?id=' . $user->id . '"',
+                    'action="http://localhost/users/edit?' . $user->id . '"',
                     $html
                 );
             } else {
                 $this->assertStringContainsString('edit-' . $user->id, $html);
-                $this->assertStringContainsString('action="http://localhost/users/edit?id=' . $user->id . '"', $html);
+                $this->assertStringContainsString('action="http://localhost/users/edit?' . $user->id . '"', $html);
             }
         }
     }